refactor: Move Hooks to the api package
Move Hooks to the api package so that the connectivity package can
use the Hooks without creating a circular dependency once it moves
to cilium/cilium repo.

Ref: https://github.com/cilium/design-cfps/pull/9
